+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Changes:
 * [BUGFIX]
 
 * [FEATURE] Add tls-min-version and tls-max-version config options
+* [FEATURE] Add an optional tls-server-name config option used to verify the server hostname
+* [BUGFIX] Fix a race condition due to which targets could use custom CA from different config sections
 
 ## 0.19.0 / 2026-03-18